+++ CMakeLists.txt
@@ -2,7 +2,12 @@ cmake_minimum_required(VERSION 3.12)
 project(gorillagame)
 
 set(CMAKE_CXX_STANDARD 17)
-if (WIN32)
+
+if(NOT APPLE)
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fsanitize=address,leak")
+endif ()
+
+    if (WIN32)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /W4")
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /wd4100")
     # -Wno-unknown-attributes gibt es auf MSVC nicht.
@@ -55,7 +60,7 @@ add_executable(${PROJECT_NAME}
         src/Map/RoomManager.cpp
         src/Map/RoomManager.h
         src/Map/Floor.cpp
-        src/Map/Floor.h src/lib/ph.cpp src/lib/ph.h)
+        src/Map/Floor.h src/lib/ph.cpp src/lib/ph.h src/lib/BulletRingBuffer.cpp src/lib/BulletRingBuffer.h)
 if (NOT WIN32)
     include(FindPkgConfig)
 endif ()
